requirements.txt
Django==4.1.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
celery==5.2.7
redis==4.3.4
django-extensions==3.2.1
dataclasses-json==0.5.7
djangorestframework-camel-case==1.3.0
requests==2.28.1
boto3==1.24.81
botocore==1.27.81
gunicorn==20.1.0
django-simple-history==3.0.0
drf-nested-routers==0.93.4